Fix instanceof check for HTMLElement
Thanks to the warning from @TheSpyder!

The issue is that `element instanceof HTMLElement` only works if
`element` is an instance of the current window's `HTMLElement`
constructor, not `HTMLElement` from another window (or iframe).

Using a variant of https://stackoverflow.com/a/59182283/20371 to do the
check, we can ensure that the `instanceof` check is done against the
element's own window's `HTMLElement` constructor.